+++ packages/destination-actions/src/destinations/amplitude/events-constants.ts
